タイトル画面のコンポーネントをまとめる
