+++ frontend/src/consts.ts
@@ -1,3 +1,1 @@
-export const title = "カタチモンスター";
-export const titleMessage = "タッチしてスタート";
 export const maxWidth = "sm";
